Move search page to component. Rename user-search to search
