+++ src/app/components/steps/SearchStep/helpers.js
@@ -1,0 +1,5 @@
+export const removeEmptyValues = (obj) => {
+  return Object.fromEntries(
+    Object.entries(obj).filter(([, v]) => v != null && v !== '')
+  );
+};
