+++ src/app/search/page.js
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+import Layout from '../components/Layout';
+import SearchStep from '../components/steps/SearchStep';
+
+const Search = () => {
+  return (
+    <Layout>
+      <SearchStep />
+    </Layout>
+  );
+};
+
+export default Search;
